Fix (hack) to prevent a ClassNotFoundException
When Arquillian iterates through to figure out the observers, it
has to iterate through each method in the class. When it does this,
*IF* FacesContext cannot be resolved, it will error out with a
ClassNotFoundException.

The circumstances of this error are the tests within a JSF project
that itself does not use JSF. For example, a test that initialises
and uses JPA within a test project